requirements.txt
# generated by showmereqs
# python version: 3.13.0

# package without version
# [package]                [import name]
Flask                    # flask
google-auth-oauthlib     # google_auth_oauthlib
google_api_python_client # googleapiclient
protobuf                 # google
waitress                 

